feat: extract interfaces to separate `webapps-schema` module
