+++ parent/pom.xml
@@ -555,6 +555,12 @@
         <version>${project.version}</version>
       </dependency>
 
+      <dependency>
+        <groupId>io.camunda</groupId>
+        <artifactId>webapps-schema</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
       <!-- Operate Modules -->
       <dependency>
         <groupId>io.camunda</groupId>
