+++ pom.xml
@@ -25,6 +25,7 @@
     <module>tasklist</module>
     <module>identity</module>
     <module>webapps-common</module>
+    <module>webapps-schema</module>
     <module>authentication</module>
     <module>search</module>
     <module>service</module>
