+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/entities/AuthorizationEntity.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.exporter.entities;
 
+import io.camunda.webapps.schema.entities.ExporterEntity;
 import io.camunda.zeebe.protocol.record.value.AuthorizationOwnerType;
 import io.camunda.zeebe.protocol.record.value.AuthorizationResourceType;
 import io.camunda.zeebe.protocol.record.value.PermissionType;
