+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/entities/TestExporterEntity.java
@@ -7,6 +7,8 @@
  */
 package io.camunda.exporter.entities;
 
+import io.camunda.webapps.schema.entities.ExporterEntity;
+
 public class TestExporterEntity implements ExporterEntity<TestExporterEntity> {
   private String id;
