+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/schema/TestUtil.java
@@ -10,8 +10,8 @@ package io.camunda.exporter.schema;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import io.camunda.exporter.schema.descriptors.IndexDescriptor;
-import io.camunda.exporter.schema.descriptors.IndexTemplateDescriptor;
+import io.camunda.webapps.schema.descriptors.IndexDescriptor;
+import io.camunda.webapps.schema.descriptors.IndexTemplateDescriptor;
 import java.util.List;
 
 public final class TestUtil {
